controllers/user/me: Move `update_email_notifications()` to dedicated module
